:muscle: Move `core/proxy` into `proxy`
