+++ denops/gin/main.ts
@@ -3,7 +3,7 @@ import type { Denops } from "https://deno.land/x/denops_std@v4.0.0/mod.ts";
 import { main as mainAction } from "./core/action/main.ts";
 import { main as mainBare } from "./core/bare/main.ts";
 import { main as mainBuffer } from "./core/buffer/main.ts";
-import { main as mainProxy } from "./core/proxy/main.ts";
+import { main as mainProxy } from "./proxy/main.ts";
 import { main as mainUtil } from "./util/main.ts";
 
 import { main as mainBranch } from "./feat/branch/main.ts";
