+++ denops/gin/proxy/editor.ts
